Move transform class from presets.py to transforms.py
