+++ examples/omnivore/data/transforms.py
@@ -9,11 +9,14 @@
 import math
 from typing import Tuple
 
+import numpy as np
+
 import torch
+import torchvision.transforms as T
 from torch import Tensor
 
 
-class RandomMixup(torch.nn.Module):
+class RandomMixup(torch.torch.nn.Module):
     """Randomly apply Mixup to the provided batch and targets.
     The class implements the data augmentations as described in the paper
     `"mixup: Beyond Empirical Risk Minimization" <https://arxiv.org/abs/1710.09412>`_.
@@ -106,7 +109,7 @@ class RandomMixup(torch.nn.Module):
         return s
 
 
-class RandomCutmix(torch.nn.Module):
+class RandomCutmix(torch.torch.nn.Module):
     """Randomly apply Cutmix to the provided batch and targets.
     The class implements the data augmentations as described in the paper
     `"CutMix: Regularization Strategy to Train Strong Classifiers with Localizable Features"
@@ -210,3 +213,153 @@ class RandomCutmix(torch.nn.Module):
             f")"
         )
         return s
+
+
+class Unsqueeze(torch.torch.nn.Module):
+    def __init__(self, pos=0):
+        super().__init__()
+        self.pos = pos
+
+    def forward(self, x):
+        return x.unsqueeze(self.pos)
+
+
+class ConvertTCHWtoCTHW(torch.nn.Module):
+    """Convert tensor from (T, C, H, W) to (C, T, H, W)"""
+
+    def forward(self, vid: torch.Tensor) -> torch.Tensor:
+        return vid.permute(1, 0, 2, 3)
+
+
+class ColorJitter3d(T.ColorJitter):
+    def forward(self, img):
+        assert isinstance(img, torch.Tensor)
+        img[:3, :, :] = super().forward(img[:3, :, :])
+        return img
+
+
+# From original implementation
+# https://www.internalfb.com/code/fbsource/[f1a98f41bcce7ee621f0248a6e0235a3e3dea628]/
+# fbcode/deeplearning/projects/omnivore/vissl/data/ssl_transforms/depth_T.py?lines=13
+class DropChannels(torch.nn.Module):
+    """
+    Drops Channels with predefined probability values.
+    Pads the dropped channels with `pad_value`.
+    Channels can be tied using `tie_channels`
+    For example, for RGBD input, RGB can be tied by using `tie_channels=[0,1,2]`.
+    In this case, channels [0,1,2] will be dropped all at once or not at all.
+    Assumes input is of the form CxHxW or TxCxHxW
+    """
+
+    def __init__(
+        self, channel_probs, fill_values, tie_channels=None, all_channel_drop=False
+    ):
+        """
+        channel_probs: List of probabilities
+        fill_values: List of values to fill the dropped channels with
+        tie_channels: List of indices. Tie dropping of certain channels.
+        all_channel_drop: Bool variable to prevent cases where all channels are dropped.
+        """
+        super().__init__()
+        channel_probs = np.array(channel_probs, dtype=np.float32)
+
+        self.channel_probs = channel_probs
+        self.fill_values = fill_values
+        self.tie_channels = tie_channels
+        self.all_channel_drop = all_channel_drop
+
+        if tie_channels is not None:
+            tie_probs = [channel_probs[x] for x in tie_channels]
+            assert len(set(tie_probs)) == 1, "All tie_channel probs must be equal"
+
+    def forward(self, x):
+        assert isinstance(x, torch.Tensor)
+        if x.ndim == 3:
+            # CxHxW
+            num_channels = x.shape[0]
+            channel_index = 0
+        elif x.ndim == 4:
+            # TxCxHxW
+            num_channels = x.shape[1]
+            channel_index = 1
+        else:
+            raise ValueError(f"Unexpected number of dims {x.ndim}. Expected 3 or 4.")
+
+        assert num_channels == len(
+            self.channel_probs
+        ), f"channel_probs is {len(self.channel_probs)} but got {num_channels} channels"
+
+        to_drop = [
+            np.random.random() < self.channel_probs[c] for c in range(num_channels)
+        ]
+        if self.tie_channels is not None:
+            first_drop = to_drop[self.tie_channels[0]]
+            for idx in self.tie_channels[1:]:
+                to_drop[idx] = first_drop
+
+        if all(to_drop) and self.all_channel_drop is False:
+            # all channels will be dropped, prevent it
+            to_drop = [False for _ in range(num_channels)]
+
+        for c in range(num_channels):
+            if not to_drop[c]:
+                continue
+            if channel_index == 0:
+                x[c, ...] = self.fill_values[c]
+            elif channel_index == 1:
+                x[:, c, ...] = self.fill_values[c]
+            else:
+                raise NotImplementedError()
+        return x
+
+
+# From original implementation:
+# https://github.com/facebookresearch/omnivore/blob/main/omnivore/T.py#L16
+class DepthNorm(torch.nn.Module):
+    """
+    Normalize the depth channel: in an RGBD input of shape (4, H, W),
+    only the last channel is modified.
+    The depth channel is also clamped at 0.0. The Midas depth prediction
+    model outputs inverse depth maps - negative values correspond
+    to distances far away so can be clamped at 0.0
+    """
+
+    def __init__(
+        self,
+        max_depth: float,
+        clamp_max_before_scale: bool = False,
+        min_depth: float = 0.01,
+    ):
+        """
+        Args:
+            max_depth (float): The max value of depth for the dataset
+            clamp_max (bool): Whether to clamp to max_depth or to divide by max_depth
+        """
+        super().__init__()
+        if max_depth < 0.0:
+            raise ValueError("max_depth must be > 0; got %.2f" % max_depth)
+        self.max_depth = max_depth
+        self.clamp_max_before_scale = clamp_max_before_scale
+        self.min_depth = min_depth
+
+    def forward(self, image: torch.Tensor):
+        c, h, w = image.shape
+        if c != 4:
+            err_msg = (
+                f"This transform is for 4 channel RGBD input only; got {image.shape}"
+            )
+            raise ValueError(err_msg)
+        color_img = image[:3, ...]  # (3, H, W)
+        depth_img = image[3:4, ...]  # (1, H, W)
+
+        # Clamp to 0.0 to prevent negative depth values
+        depth_img = depth_img.clamp(min=self.min_depth)
+
+        # divide by max_depth
+        if self.clamp_max_before_scale:
+            depth_img = depth_img.clamp(max=self.max_depth)
+
+        depth_img /= self.max_depth
+
+        img = torch.cat([color_img, depth_img], dim=0)
+        return img
